// paris-sportifs/src/app/models/team.ts
export interface Team {
  idTeam: string;
  idSoccerXML: string;
  idAPIfootball: string;
  intLoved: string | null;
  strTeam: string;
  strTeamShort: string | null;
  strAlternate: string;
  intFormedYear: string;
  strSport: string;
  strLeague: string;
  idLeague: string | null;
  strLeague2: string;
  idLeague2: string | null;
  strLeague3: string;
  idLeague3: string | null;
  strLeague4: string;
  idLeague4: string | null;
  strLeague5: string;
  idLeague5: string | null;
  strLeague6: string;
  idLeague6: string | null;
  strLeague7: string;
  idLeague7: string | null;
  strDivision: string | null;
  strManager: string;
  strStadium: string;
  strKeywords: string;
  strRSS: string;
  strStadiumThumb: string;
  strStadiumDescription: string;
  strStadiumLocation: string;
  intStadiumCapacity: string;
  strWebsite: string;
  strFacebook: string;
  strTwitter: string;
  strInstagram: string;
  strDescriptionEN: string | null;
  strDescriptionDE: string | null;
  strDescriptionFR: string | null;
  strDescriptionCN: string | null;
  strDescriptionIT: string | null;
  strDescriptionJP: string | null;
  strDescriptionRU: string | null;
  strDescriptionES: string | null;
  strDescriptionPT: string | null;
  strDescriptionSE: string | null;
  strDescriptionNL: string | null;
  strDescriptionHU: string | null;
  strDescriptionNO: string | null;
  strDescriptionIL: string | null;
  strDescriptionPL: string | null;
  strKitColour1: string;
  strKitColour2: string;
  strKitColour3: string;
  strGender: string;
  strCountry: string;
  strTeamBadge: string;
  strTeamJersey: string;
  strTeamLogo: string;
  strTeamFanart1: string;
  strTeamFanart2: string;
  strTeamFanart3: string;
  strTeamFanart4: string;
  strTeamBanner: string;
  strYoutube: string;
  strLocked: string;
}